+++ mirte_base_control/src/my_robot_base.cpp
@@ -1,6 +1,6 @@
 #include <chrono>
 #include <functional>
-#include <my_robot_hw_interface.h>
+#include <mirte_base_control.hpp>
 // #include <ros/callback_queue.h>
 
 void controlLoop(MirteBaseHWInterface &hw,
